requirements.txt
psycopg2-binary
openai
databricks-sdk>=0.81.0
